refactor(react): rename the hook to __experimentalBindRenderContext
bind describes the operation (create a render context and bind it to the
caller's bridge objects); page vocabulary is gone. Files follow:
multi-root-render-context.ts / bound-root.ts.
