+++ .changeset/reactlynx-create-root.md
@@ -6,4 +6,4 @@
 
 Support multiple pages sharing one background JS context, behind the `experimental_multiRootRenderContext` build option (`__MULTI_ROOT_RENDER_CONTEXT__` define, off by default).
 
-Each page gets its own root with isolated render state and native channels. Pages keep using the classic `root.render`; a per-page glue call (`root.__experimentalBootstrapPage?.(...)`, installed on `root` only in multi-page builds) binds the shared runtime to the loading page, with no new exported API. Normal builds strip the whole mechanism.
+Each page gets its own root with isolated render state and native channels. Pages keep using the classic `root.render`; a per-page glue call (`root.__experimentalBindRenderContext?.(...)`, installed on `root` only in multi-page builds) binds the shared runtime to the loading page, with no new exported API. Normal builds strip the whole mechanism.
