+++ packages/react/runtime/__test__/snapshot/multiRootRenderContext.test.jsx
@@ -10,8 +10,8 @@ import { useState } from '../../src/index';
 import { root } from '../../src/lynx-api';
 import { __root } from '../../src/root';
 import { defaultRootContext, switchRootContext } from '../../src/root-context';
-import { ReactLynxRoot } from '../../src/multi-page';
-import '../../src/multi-page';
+import { ReactLynxRoot } from '../../src/multi-root-render-context';
+import '../../src/multi-root-render-context';
 import { backgroundSnapshotInstanceManager, setupPage } from '../../src/snapshot';
 import { replaceCommitHook } from '../../src/snapshot/lifecycle/patch/commit';
 import { initGlobalSnapshotPatch } from '../../src/snapshot/lifecycle/patch/snapshotPatch';
@@ -59,14 +59,14 @@ describe('root-cause: default singleton shares one container', () => {
 
 describe('bootstrap hook: classic root.render renders the bootstrapped page', () => {
   afterEach(() => {
-    root.__experimentalBootstrapPage();
+    root.__experimentalBindRenderContext();
   });
 
   it('delegates root.render to the bootstrapped root, leaving __root untouched', () => {
     globalEnvManager.switchToBackground();
 
     const A = () => <text>{'A'}</text>;
-    const pageA = root.__experimentalBootstrapPage({});
+    const pageA = root.__experimentalBindRenderContext({});
     root.render(<A />);
 
     expect(textOf(pageA._container)).toEqual(['A']);
@@ -79,9 +79,9 @@ describe('bootstrap hook: classic root.render renders the bootstrapped page', ()
     const A = () => <text>{'A'}</text>;
     const B = () => <text>{'B'}</text>;
 
-    const pageA = root.__experimentalBootstrapPage({});
+    const pageA = root.__experimentalBindRenderContext({});
     root.render(<A />);
-    const pageB = root.__experimentalBootstrapPage({});
+    const pageB = root.__experimentalBindRenderContext({});
     root.render(<B />);
 
     expect(pageA._container).not.toBe(pageB._container);
@@ -93,8 +93,8 @@ describe('bootstrap hook: classic root.render renders the bootstrapped page', ()
     globalEnvManager.switchToBackground();
 
     const A = () => <text>{'A'}</text>;
-    root.__experimentalBootstrapPage({});
-    root.__experimentalBootstrapPage();
+    root.__experimentalBindRenderContext({});
+    root.__experimentalBindRenderContext();
     root.render(<A />);
 
     expect(textOf(__root)).toEqual(['A']);
@@ -112,7 +112,7 @@ describe('bootstrap hook: classic root.render renders the bootstrapped page', ()
     const old = globalThis.__FIRST_SCREEN_SYNC_TIMING__;
     try {
       globalThis.__FIRST_SCREEN_SYNC_TIMING__ = 'jsReady';
-      root.__experimentalBootstrapPage({ lynx: lynxA });
+      root.__experimentalBindRenderContext({ lynx: lynxA });
       root.render(<text>{'A'}</text>);
       const handshakes = callA.mock.calls.filter(c => c[0] === 'rLynxFirstScreenSyncReady');
       expect(handshakes).toHaveLength(1);
