+++ packages/react/runtime/src/bound-root.ts
@@ -1,0 +1,10 @@
+// Copyright 2026 The Lynx Authors. All rights reserved.
+// Licensed under the Apache License Version 2.0 that can be found in the
+// LICENSE file in the root directory of this source tree.
+import type { ReactLynxRoot } from './multi-root-render-context.js';
+
+export let boundRoot: ReactLynxRoot | undefined;
+
+export function setBoundRoot(root: ReactLynxRoot | undefined): void {
+  boundRoot = root;
+}
