--- packages/react/runtime/src/page-root-ref.ts
@@ -1,10 +1,0 @@
-// Copyright 2026 The Lynx Authors. All rights reserved.
-// Licensed under the Apache License Version 2.0 that can be found in the
-// LICENSE file in the root directory of this source tree.
-import type { ReactLynxRoot } from './multi-page.js';
-
-export let bootstrappedRoot: ReactLynxRoot | undefined;
-
-export function setBootstrappedRoot(root: ReactLynxRoot | undefined): void {
-  bootstrappedRoot = root;
-}
